+++ test/fixtures/pkg-with-config/jetpack.config.mjs
@@ -2,10 +2,5 @@ export default {
   port: 1234,
   entry: './app/client',
   title: 'testing',
-  verbose: true,
-  css: {
-    features: {
-      'nesting-rules': true
-    }
-  }
+  verbose: true
 }