+++ test/fixtures/pkg-with-everything/jetpack.config.mjs
@@ -3,11 +3,6 @@ import rspackConfigForTests from '../../helpers/rspackConfigForTests.js'
 export default {
   minify: false,
   chunkLoadRetry: true,
-  css: {
-    features: {
-      'nesting-rules': true
-    }
-  },
   rspack(config) {
     rspackConfigForTests(config)
   }